HBASE-2054 memstore size 0 is >= than blocking -2.0g size
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/branches/0.20@892651 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/HTableDescriptor.java
@@ -412,9 +412,9 @@ ISerializable {
   /**
    * @param memstoreFlushSize memory cache flush size for each hregion
    */
-  public void setMemStoreFlushSize(int memstoreFlushSize) {
+  public void setMemStoreFlushSize(long memstoreFlushSize) {
     setValue(MEMSTORE_FLUSHSIZE_KEY,
-      Bytes.toBytes(Integer.toString(memstoreFlushSize)));
+      Bytes.toBytes(Long.toString(memstoreFlushSize)));
   }
 
   /**
